server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["shadowciaw <sqli@ualberta.ca>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.3"
python-dotenv = "^0.19.2"
Flask-PyMongo = "^2.3.0"
Werkzeug = "^2.0.3"
pymongo = {version = "^4.0.2", extras = ["srv"]}
requests = "^2.27.1"
Flask-API = "^3.0.post1"
pytest = "^7.1.1"
unittest = "^0.0"
mongomock = "^4.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
